5.3.6 Requires validator to submit previous proposed state root
This commit modifies the propose state root function to take an
additional argument: the state root proposed for the previous epoch.
This helps prevent the case where two validators submit a transaction
to propose the same state root right after each other (and generally
informs the validator if their understanding of the state history
is correct). Also adds epoch key when storing the block when a
state root is proposed. This is necessary to prevent griefing where
a validator submits a valid state root for the wrong epoch.